requirements.txt
tensorflow==2.15.0
kafka-python==2.0.2
numpy==1.26.4
fastapi==0.115.0
uvicorn==0.30.6
pyyaml==6.0.2
pytest==8.3.3
python-jose[cryptography]==3.3.0
prometheus-client==0.20.0
flake8==7.1.1